Suggested changes for the README.md
There is no guide on how to fork or download the codebase from this GitHub repository to execute main.py from Spyder yet. @fazamahezs please add and review the following guide to ensure it is sufficient for a new user to successfully run the model. Fix this by tomorrow so that we can ask @yumnakarimah to try it if she is available.